Comments Logic Implemented

+++ week3/starter/pokemonservices/comment/src/main/java/com/fiipractic/comment/model/Comment.java
@@ -26,7 +26,7 @@ public class Comment {
     private double rating;
 
     @Column(name="pokemon_id")
-    private Long pokemonId;
+    private Integer pokemonId;
 
     @Column(name="comment_description")
     private String commentDescription;
